// supabase/functions/_shared/engine/types.ts
export const SUITS = ['pik', 'karo', 'herc', 'tref'] as const
export type Suit = (typeof SUITS)[number]

// Rang u boji, slabije → jače (7 najslabiji, A najjači)
export const RANKS = ['7', '8', '9', '10', 'J', 'Q', 'K', 'A'] as const
export type Rank = (typeof RANKS)[number]

export interface Card {
  suit: Suit
  rank: Rank
}